Refactor: Modularize agent GUI and separate BasketballAgent class into its own file
